+++ bigml/tests/test_05_compare_predictions.py
@@ -590,3 +590,48 @@ class TestComparePrediction(object):
             prediction_create.the_logistic_prediction_is(self, example[6])
             prediction_compare.i_create_a_local_prediction(self, example[5])
             prediction_compare.the_local_prediction_is(self, example[6])
+
+    def test_scenario13(self):
+        """
+            Scenario: Successfully comparing predictions:
+                Given I create a data source uploading a "<data>" file
+                And I wait until the source is ready less than <time_1> secs
+                And I create a dataset
+                And I wait until the dataset is ready less than <time_2> secs
+                And I create a model
+                And I wait until the model is ready less than <time_3> secs
+                And I create a local model
+                When I create a prediction for "<data_input>"
+                Then the prediction for "<objective>" is "<prediction>"
+                And I create a local prediction for "<data_input>"
+                Then the local prediction is "<prediction>"
+
+                Examples:
+                | data             | time_1  | time_2 | time_3 | data_input                             | objective | prediction  |
+
+        """
+        examples = [
+            ['data/iris.csv', '10', '10', '10', '{"petal width": 0.5}', '000004', 'Iris-setosa', "tmp/my_model.json", "my_test"],
+            ['data/iris.csv', '10', '10', '10', '{"petal length": 6, "petal width": 2}', '000004', 'Iris-virginica', "tmp/my_model.json", "my_test"],
+            ['data/iris.csv', '10', '10', '10', '{"petal length": 4, "petal width": 1.5}', '000004', 'Iris-versicolor', "tmp/my_model.json", "my_test"],
+            ['data/iris_sp_chars.csv', '10', '10', '10', '{"pétal.length": 4, "pétal&width\u0000": 1.5}', '000004', 'Iris-versicolor', "tmp/my_model.json", "my_test"]]
+        show_doc(self.test_scenario13, examples)
+        for example in examples:
+            print "\nTesting with:\n", example
+            source_create.i_upload_a_file(self, example[0])
+            source_create.the_source_is_finished(self, example[1])
+            dataset_create.i_create_a_dataset(self)
+            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
+            args = '{"tags": ["%s"]}' % example[8]
+            model_create.i_create_a_model_with(self, data=args)
+            model_create.the_model_is_finished_in_less_than(self, example[3])
+            model_create.i_export_model(self, example[7])
+            prediction_compare.i_create_a_local_model_from_file(self, example[7])
+            prediction_create.i_create_a_prediction(self, example[4])
+            prediction_create.the_prediction_is(self, example[5], example[6])
+            prediction_compare.i_create_a_local_prediction(self, example[4])
+            prediction_compare.the_local_prediction_is(self, example[6])
+            model_create.i_export_tags_model(self, example[7], example[8])
+            prediction_compare.i_create_a_local_model_from_file(self, example[7])
+            prediction_compare.i_create_a_local_prediction(self, example[4])
+            prediction_compare.the_local_prediction_is(self, example[6])